+++ app/src/main/java/com/example/eddyapp/data/model/ClientBatteryResponse.kt
@@ -2,9 +2,9 @@ package com.example.eddyapp.data.model
 
 data class ClientBatteryResponse(
     val status: String,
-    val charge: Int,
+    val charge: Float,
     val charging: Boolean,
-    val remaining_time: Int,
+    val remaining_time: Float,
 
     // optional field message: String in case of error
     val message: String? = null
